requirements_dev.txt
# quantify

pip
bump2version==1.0.1
wheel==0.33.6
watchdog==0.9.0
coverage
sphinx>=2,<4 # jupyter_sphinx requires sphinx>=2, <4 avoids a sphinx warning from jupyter_sphinx
sphinx-rtd-theme==0.5.0rc2
sphinxcontrib-bibtex
jupyter_sphinx>=0.3.2
sphinx-jsonschema>=1.15
sphinx-autobuild
sphinx-togglebutton
sphinx-autodoc-typehints
twine==1.14.0
pytest
pytest-runner
pytest-cov
pytest-xdist # run parallel tests
pygments==2.6.1

black
pylint
pre-commit
pre-commit-hooks # common hooks for git repos
isort # sort python imports

scanpydoc>=0.7.1 # ensures latest typing_extensions is used in sphinx (https://github.com/theislab/scanpydoc/issues/31)
docutils<0.17 # pinned due to issue #175
rich[jupyter]
jupytext
qcodes>=0.28.0 # document Parameter attributes

# quantify-core only

scikit-learn
scikit-optimize
quantify-scheduler # used to generate schedule example in the dataset design docs
